requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
engineering-notation==0.10.0
idna==3.6
pillow==10.2.0
requests==2.31.0
tk==0.1.0
tk-tools==0.16.0
urllib3==2.2.0
